Shroud, ProjectedCellTouchStateLayer maintain an array of touched UVs.
In `Shroud.Tick` the complete map region used to be iterated to search
for touched cells in the `touched` cell layer.

The TouchState cell layer maintains an array of cells that are touched
avoiding the need to iterate the complete map region to find them.

Tick time improvement of at least 50%.

'Fastest' way to process `touched` cells.

On average a maximum of to 30-50% of cells are estimated to be marked
as touched throughout a busy multiplayer game.

At the memory cost of an additional - internal - cell layer.

`Shroud.Tick` no longer processes touched cells in map order. This
does not seem to have an impact (on Sync).

Slower at marking cells "touched". Faster at "finding" touched cells.

Slower once over 50% of cells are touched because a check
needs to be performed to see if a cell already was marked before.
Perhaps above the layer should fall back on past behavior of
iterating the region entries.

No map region iterations when nothing is touched.

Avoids the need for Shroud optimization based on spatial separation.

Can potentially also be used in ShroudRenderer.